GUI: follow up to #191

+++ src/gui/gui.h
@@ -417,7 +417,7 @@ class FurnaceGUI {
 
   double dpiScale;
 
-  int aboutScroll, aboutSin;
+  double aboutScroll, aboutSin;
   float aboutHue;
 
   ImFont* mainFont;
